fix: typescript interfaces in comment/post/vote
