+++ packages/frontend/src/components/comment/Comment.tsx
@@ -14,9 +14,9 @@ import {
     useMenuState,
 } from '@szhsin/react-menu'
 import { useMediaQuery } from '@uidotdev/usehooks'
-import { CommentInfo, CommentStatus } from '../../types'
 import CommentDeleteDialog from './CommentDeleteDialog'
 import CommentReportDialog from './CommentReportDialog'
+import { CommentInfo, CommentStatus } from '@/types/Comments'
 
 interface CommentProps extends CommentInfo {
     onCloseAnimation?: () => void
