+++ packages/frontend/src/components/comment/CommentNotification.tsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom'
 import CloseIcon from '@/assets/close.svg'
 import PostIcon from '@/assets/post.svg'
 import useCommentEvents from '@/hooks/useCommentEvents'
-import { CommentMsg } from '@/types'
+import { CommentMsg } from '@/types/Comments'
 
 interface CommentNotificationsProps {
     postId: string
