--- packages/frontend/src/pages/Profile/History/DTO/Post.ts
@@ -1,11 +1,0 @@
-export class Post {
-    constructor(
-        public id: string,
-        public epochKey: string,
-        public publishedAt: number,
-        public content: string,
-        public voteSum: number,
-        public date: string,
-        public url: string,
-    ) {}
-}
