--- packages/frontend/src/pages/Profile/History/DTO/Vote.ts
@@ -1,13 +1,0 @@
-import { VoteType } from '../types'
-
-export class Vote {
-    constructor(
-        public epochKey: string,
-        public publishedAt: number,
-        public content: string,
-        public voteSum: number,
-        public date: string,
-        public url: string,
-        public type: VoteType,
-    ) {}
-}
