+++ packages/frontend/src/pages/Profile/History/services/CommentService.ts
@@ -1,11 +1,10 @@
 import { UserState } from '@unirep/core'
 import dayjs from 'dayjs'
 
-import { RelayRawComment } from '../../../../types/api'
 import { fetchCommentsByEpochKeys } from '../../../../utils/api'
-import { Comment } from '../DTO/Comment'
 import { ActiveFilter } from '../types'
 import { fetchAllByEpochKeysInBatches } from '../utils'
+import { CommentHistoryMetaData, RelayRawComment } from '@/types/Comments'
 
 export class CommentService {
     async fetchCommentHistoryByUserState(userState: UserState) {
@@ -15,13 +14,12 @@ export class CommentService {
             chunkSize,
             fetchCommentsByEpochKeys,
         )
-        console.log(batchedRawComments)
         return batchedRawComments
             .map(this.parseRelayRawCommentsToComments.bind(this))
             .flat(2)
     }
 
-    sortComments(comments: Comment[], activeFilter: ActiveFilter) {
+    sortComments(comments: CommentHistoryMetaData[], activeFilter: ActiveFilter) {
         switch (activeFilter) {
             case ActiveFilter.DateAsc: {
                 return this.sortCommentsByDateAsc(comments)
@@ -38,24 +36,24 @@ export class CommentService {
         }
     }
 
-    private sortCommentsByDateAsc(comments: Comment[]) {
+    private sortCommentsByDateAsc(comments: CommentHistoryMetaData[]) {
         return comments.sort((a, b) => b.publishedAt - a.publishedAt)
     }
 
-    private sortCommentsByDateDesc(comments: Comment[]) {
+    private sortCommentsByDateDesc(comments: CommentHistoryMetaData[]) {
         return comments.sort((a, b) => a.publishedAt - b.publishedAt)
     }
 
-    private sortCommentsByPopularityAsc(comments: Comment[]) {
+    private sortCommentsByPopularityAsc(comments: CommentHistoryMetaData[]) {
         return comments.sort((a, b) => b.voteSum - a.voteSum)
     }
 
     private parseRelayRawCommentsToComments(
         relayRawComments: RelayRawComment[],
-    ): Comment[] {
+    ): CommentHistoryMetaData[] {
         return relayRawComments.map((relayRawComment) => {
             const publishedAt = parseInt(relayRawComment.publishedAt)
-            return new Comment(
+            return new CommentHistoryMetaData(
                 relayRawComment.commentId,
                 relayRawComment.postId,
                 relayRawComment.epochKey,
