+++ packages/frontend/src/pages/Profile/History/utils.ts
@@ -39,6 +39,5 @@ export async function fetchAllByEpochKeysInBatches<T, U>(
             fetchCall({ epochKeys: epochKeyChunk, ...restFetchCallParams }),
         ),
     )
-    console.log(batchedRawPosts)
     return batchedRawPosts
 }
