+++ packages/frontend/src/store/storeTypes.ts
@@ -1,4 +1,4 @@
-import { PostInfo } from '@/types'
+import { PostInfo } from "@/types/Post"
 
 export interface StoreState {
     posts: PostInfo[]
